update crossover validation logic
Still not perfect - we may have to come up with a factor other than
rotation to explain degree of crossiness: being crossed on the center
two arrows is technically 180, but extending the crossed foot further to
the next up/down arrow feels more crossed even though it reduces
rotation to 153.  Extending the crossed foot to the far left/right arrow
feels *much* more crossed even though it is still just 180.